@@ -26,6 +26,7 @@
   },
   "devDependencies": {
     "@balena/lint": "5.4.1",
+    "@electron/remote": "^2.0.10",
     "@fortawesome/fontawesome-free": "5.14.0",
     "@svgr/webpack": "5.5.0",
     "@types/debug": "4.1.4",
@@ -41,26 +42,26 @@
     "css-loader": "5.0.1",
     "dbus-next": "0.10.2",
     "debug": "4.3.1",
-    "electron": "12.0.2",
+    "electron": "22.3.18",
     "file-loader": "6.2.0",
-    "html-webpack-plugin": "5.0.0-beta.1",
+    "html-webpack-plugin": "5.5.3",
     "husky": "4.2.5",
     "lint-staged": "10.2.11",
     "lockfile": "1.0.4",
-    "lodash": "4.17.20",
+    "lodash": "4.17.21",
     "mini-css-extract-plugin": "1.3.3",
     "pretty-bytes": "5.2.0",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "rendition": "20.10.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "rendition": "34.16.4",
     "style-loader": "2.0.0",
     "sys-class-rgb-led": "3.0.1",
-    "ts-json-schema-generator": "0.82.0",
-    "ts-loader": "8.1.0",
-    "ts-node": "9.0.0",
-    "typescript": "4.0.3",
-    "webpack": "5.30.0",
-    "webpack-cli": "4.6.0",
+    "ts-json-schema-generator": "2.4.0",
+    "ts-loader": "9.4.4",
+    "ts-node": "10.9.1",
+    "typescript": "5.1.6",
+    "webpack": "5.88.2",
+    "webpack-cli": "5.1.4",
     "x11": "2.3.0"
   },
   "versionist": {
